Install carthage using .pkg from release.
Carthage was being installed using homebrew. Due to homebrew
not being able to always have the latest version of carthage
we now install it using the .pkg from carthage's github
release.

script/build, script/test and script/update will check
if the installed version of carthage matches the
required version specified in script/.env. This is done
to force developers to update carthage every time
there is a new version.