Splits bootstrap hooks in two phases.
Now the bootstrap hooks are split in two phased. Before and
after installing dependencies. This way the extension point
is more flexible.

This is a breaking change because the
script/script_hooks/bootstrap file is no longer executed, is
sourced. If defined the bootstrap_before_install_hooks and
bootstrap_after_install_hooks functions are executed during
the bootstrap process in the appropiate places.